+++ pkg/certificates/certificates_client/client.go
@@ -1,7 +1,7 @@
 // Copyright (C) 2019 Storj Labs, Inc.
 // See LICENSE for copying information.
 
-package client
+package certificates_client
 
 import (
 	"context"
